pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerautomate-cli"
version = "0.1.0"
description = "CLI interface for Microsoft Power Automate Management API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Adam Bertram", email = "adbertram@gmail.com"}
]
keywords = ["power-automate", "power-platform", "flow", "cli", "api", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer[all]>=0.9.0",
    "requests>=2.31.0",
    "msal>=1.24.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "openapi-spec-validator>=0.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
powerautomate = "powerautomate_cli.main:main"

[project.urls]
"Homepage" = "https://github.com/adbertram/powerautomate-cli"
"Bug Tracker" = "https://github.com/adbertram/powerautomate-cli/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["powerautomate_cli*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=powerautomate_cli --cov-report=term-missing"
